+++ entities/CAttackers.hpp
